two transcription models in now torchserve.
