+++ src/python/assistant/apps/torchserve/build_script.py
@@ -1,3 +1,4 @@
 import whisper
 
 whisper.load_model('tiny.en')
+whisper.load_model('base.en')
