+++ src/python/assistant/apps/torchserve/readme.md
@@ -7,7 +7,8 @@ This implements Torchserve for running the models in this project for inference.
 torch-model-archiver —-model-name “<model name>” —version 1.0 —-extra-files “./config.json”,”./utils.py” —handler “<path to handler>” —export-path model_store
 
 
-torch-model-archiver --model-name "jenkins_prompt" --version "1.0" --extra-files "./config.json" --handler "./models/jenkins_prompt.py" --export-path model_store
+torch-model-archiver --model-name "jenkins_prompt" --version "1.0" --extra-files "./config.json","./models/audio_processors/audio_transcription.py" --handler "./models/jenkins_prompt.py" --export-path model_store --force
 
+torch-model-archiver --model-name "jenkins_listen" --version "1.0" --extra-files "./config.json","./models/audio_processors/audio_transcription.py" --handler "./models/jenkins_listen.py" --export-path model_store --force
 
-torchserve --start --ts-config torchserve.config --model-store "model_store" --models jenkins_prompt=jenkins_prompt.mar
+torchserve --start --ts-config torchserve.config --model-store "model_store"
