+++ src/python/assistant/apps/torchserve/torchserve.config
@@ -11,6 +11,16 @@ models = {\
         “batchSize” : 1,\
         “maxBatchDelay” : 50,\
         “responseTimeout” : 120\
-	}\
-  }\
+	  }\
+  },\
+ jenkins_listen : {\
+    "1.0": {\
+        “defaultVersion”: true,\
+        “minWorkers”: 1,\
+        “maxWorkers” : 2,\
+        “batchSize” : 1,\
+        “maxBatchDelay” : 50,\
+        “responseTimeout” : 120\
+	  }\
+ }\
 }\
